Move some common fields and code to AbstractAggregate
